+++ src/crabagent/static/assets/swimlanesDiagram-G3AALYLV-SCNgjvsd.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-23GEKE2U-DgvjZBE1.js";import{a}from"./mermaid.core-DffUkTC1.js";import"./chunk-5VM5RSS4-BB8O9h2d.js";import"./chunk-XXDRQBXY-DfaDfC-S.js";import"./chunk-VR4S4FIN-sLmahjBR.js";import"./chunk-32BRIVSS-BzJaPGMx.js";import"./channel-DVmBQtGf.js";import"./index-Bl6jM4N-.js";import"./vendor-react-BEFBu0qs.js";import"./vendor-ui-mGeIAIgk.js";import"./vendor-charts-BLKufy57.js";import"./vendor-markdown-C0jI_Wfn.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),m=o,v=r({defaultLayout:"swimlane",styles:m});export{v as diagram};